Break store out of DefaultRelationship*Cursor
